src/cf_hourly-avg-drainage/requirements.txt

 
ipykernel 
pytest 
numpy 
cognite-sdk 
pandas 
python-dotenv 
statsmodels 


